requirements.txt
streamlit
onnxruntime
Pillow
numpy
pandas
opencv-python 